+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 [Squint](https://github.com/squint-cljs/squint): ClojureScript syntax to JavaScript compiler
 
-## Unreleased
+## 0.0.13
 
 - Fix compilation of empty list
 - Fix `and` and `or` without arguments
-- #308: `doseq` in return position emits invalid code
+- [#308](https://github.com/squint-cljs/squint/issues/308): `doseq` in return position emits invalid code
 
 ## 0.0.12